Test profile

+++ src/main/java/com/teamproject/trackers/biz/followSubscribeLike/FollowRepository.java
@@ -19,7 +19,7 @@ public interface FollowRepository extends JpaRepository<FollowVO, FollowId> {
 	
 	// follow 가져오기
 	@Query(value = "SELECT 1 FROM follow f WHERE f.to_id = (SELECT u.id FROM user u WHERE u.url = :url) AND f.from_id = (SELECT u.id FROM user u WHERE u.id = :id)",nativeQuery = true)
-	public Long followT(@Param("url") String url, @Param("id") long id);
+	public int followT(@Param("url") String url, @Param("id") long id);
 	
 	// 팔로우 정보 수정 -구독한 사람의 정보를 제거하는거(팔로우 취소)
 	//@Query(value="DELETE FROM follow WHERE from_id = :fromid  and to_id = :toid", nativeQuery = true)
